requirements.txt
Flask>=2.0.0
scikit-learn
pandas
numpy
joblib
python-whois
tldextract
requests
python-dateutil